+++ docs/Whats-New.md
@@ -92,6 +92,7 @@ Vinifera fixes:
 - Fix a bug where only local variables up to 50 would be used (by ZivDero)
 - Fix a bug where you'd hear "Construction complete" after placing a building, instead of when it's ready (by ZivDero)
 - Fix a bug where AI vehicles got stuck on War Factory if attacked while exiting (by Rampastring, ZivDero)
+- Fix a bug where you could tote a `Totable=no` unit by force-moving onto it (by ZivDero)
 
 
 Vanilla fixes: